Module mithril::atms

Ad-Hoc Threshold Multisignatures

Proof-of-Stake Sidechains https://eprint.iacr.org/2018/1239.pdf

The implementation in this module is parameterized by the underlying signature scheme, which will define its own type of key and signature. All we need here is to be able to verify signatures against keys.

Structs

Asig

An Aggregated Signature

Avk

Aggregation of Keys

MTValue

The values that are represented in the Merkle Tree.

Enums

VerifyFailure

Traits

VerifySig

The implementation relies on an underlying signature scheme All that we need from it is the ability to verify that a signature is valid for a given aggregation of keys