Trait mithril::atms::Atms

pub trait Atms: Sized + Debug {
    type PreCheckedPK: Copy;
    type CheckedPK;
    type SIG;
    fn check_key(proof: &Self::PreCheckedPK) -> Option<Self::CheckedPK>;

    fn verify(msg: &[u8], pk: &Self::CheckedPK, sig: &Self::SIG) -> bool;
}

The implementation relies on an underlying public signature with a proof of possession. We need the ability to verify the signature. Moreover, it is common that multi signatures schemes require the verification of correctness of the public key, and therefore we require also associate a check_key function.

Associated Types

type PreCheckedPK: Copy

Public key before verification of correctness.

type CheckedPK

Public key after verification of correctness.

type SIG

Associated type that defines the Signature.

Required methods

fn check_key(proof: &Self::PreCheckedPK) -> Option<Self::CheckedPK>

Verify the correctness of the public key. If it is valid, return the public key (without the proof of correctness).

fn verify(msg: &[u8], pk: &Self::CheckedPK, sig: &Self::SIG) -> bool

Verify that the signature is valid for a message msg and key key