Trait mithril::merkle_tree::MTHashLeaf

pub trait MTHashLeaf<L> {
    type F: Eq + Clone + Debug;
    fn new() -> Self;

    fn zero() -> Self::F;

    fn root_bytes(h: &Self::F) -> Vec<u8>;

    fn inject(&mut self, v: &L) -> Self::F;

    fn hash_children(&mut self, left: &Self::F, right: &Self::F) -> Self::F;

    fn hash(&mut self, leaf: &[Self::F]) -> Self::F { ... }
}

This trait describes a hashing algorithm. For mithril we need (1) a way to inject stored values into the tree (2) a way to combine hashes (H_p is used for both of these in the paper)

Associated Types

type F: Eq + Clone + Debug

The output domain of the hasher.

Required methods

fn new() -> Self

Create a new hasher

fn zero() -> Self::F

This should be some “null” representative

fn root_bytes(h: &Self::F) -> Vec<u8>

How to extract hashes as bytes

fn inject(&mut self, v: &L) -> Self::F

How to map (or label) values with their hash values

fn hash_children(&mut self, left: &Self::F, right: &Self::F) -> Self::F

Combine (and hash) two hash values

Provided methods

fn hash(&mut self, leaf: &[Self::F]) -> Self::F

Hash together an arbitrary number of values, Reducing the input with zero() as the initial value and hash_children as the operation