Trait mithril::proof::Proof

pub trait Proof: Sized {
    type Env: ProverEnv;
    type Statement;
    type Relation;
    type Witness;
    type Error: Debug + Into<i64> + Display;
    fn prove(
        env: &Self::Env, 
        pk: &<Self::Env as ProverEnv>::ProvingKey, 
        rel: &Self::Relation, 
        stmt: &Self::Statement, 
        witness: Self::Witness
    ) -> Result<Self, Self::Error>;

    fn verify(
        &self, 
        env: &Self::Env, 
        vk: &<Self::Env as ProverEnv>::VerificationKey, 
        rel: &Self::Relation, 
        stmt: &Self::Statement
    ) -> Result<(), Self::Error>;
}

Implementors of Proof know how to prove that a relation of type Relation holds between values of types Statement and Witness (generally the proofs are knowledge of such a Witness)

Associated Types

type Env: ProverEnv

Context for this proof system.

type Statement

Statement to be proven.

type Relation

Relation between values of type Statement and type Witness.

type Witness

Witness of the validity of Statement.

type Error: Debug + Into<i64> + Display

Type of errors which can be output by the proof system. Into allows passing the error to the C API.

Required methods

fn prove(
env: &Self::Env,
pk: &<Self::Env as ProverEnv>::ProvingKey,
rel: &Self::Relation,
stmt: &Self::Statement,
witness: Self::Witness
) -> Result<Self, Self::Error>

Construct a proof.

fn verify(
&self,
env: &Self::Env,
vk: &<Self::Env as ProverEnv>::VerificationKey,
rel: &Self::Relation,
stmt: &Self::Statement
) -> Result<(), Self::Error>

Verify a proof.