Function mithril::stm::ev_lt_phi

pub fn ev_lt_phi(
    phi_f: f64, 
    ev: [u8; 64], 
    stake: Stake, 
    total_stake: Stake
) -> bool

The crate rug has sufficient optimizations to not require a taylor approximation with early stop. The difference between the current implementation and the one using the optimization above is around 10% faster. We perform the computations with 117 significant bits of precision, since this is enough to represent the fraction of a single lovelace. We have that 1e6 lovelace equals 1 ada, and there is 45 billion ada in circulation. Meaning there are 4.5e16 lovelace, so 1e-17 is sufficient to represent fractions of the stake distribution. In order to keep the error in the 1e-17 range, we need to carry out the computations with 34 decimal digits (in order to represent the 4.5e16 ada without any rounding errors, we need double that precision).