Struct mithril::stm::StmInitializer

pub struct StmInitializer<PE> where
    PE: PairingEngine,  { /* fields omitted */ }

Initializer for StmSigner. This is the data that is used during the key registration procedure. One the latter is finished, this instance is consumed into an StmSigner.

Implementations

impl<PE> StmInitializer<PE> where
PE: PairingEngine,
MspPk<PE>: Hash,

pub fn setup<R>(
params: StmParameters,
party_id: PartyId,
stake: Stake,
rng: &mut R
) -> Self where
R: RngCore + CryptoRng,

Builds an StmInitializer that is ready to register with the key registration service

pub fn generate_new_key<R>(&mut self, rng: &mut R) where
R: RngCore + CryptoRng,

Create a new key.

pub fn secret_key(&self) -> MspSk<PE>

Extract the secret key.

pub fn verification_key(&self) -> MspPk<PE>

Extract the verification key.

pub fn set_key(&mut self, sk: &MspSk<PE>)

Set a new pair of keys out of a secret key

pub fn set_stake(&mut self, stake: Stake)

Set the stake.

pub fn stake(&self) -> Stake

Get the stake.

pub fn party_id(&self) -> PartyId

Get the party ID.

pub fn set_params(&mut self, params: StmParameters)

Set the StmParameters.

pub fn params(&self) -> StmParameters

Get the parameters

pub fn new_signer<H>(self, closed_reg: ClosedKeyReg<PE, H>) -> StmSigner<H, PE> where
H: MTHashLeaf<MTValue<PE>> + Clone,

Build the avk for the given list of parties.

Note that if this StmInitializer was modified between the last call to register, then the resulting StmSigner may not be able to produce valid signatures.

Returns an StmSigner specialized to (1) this StmSigner’s ID and current stake (2) this StmSigner’s parameter valuation (3) the avk as built from the current registered parties (according to the registration service) (4) the current total stake (according to the registration service)