Trait mithril::atms::Atms

pub trait Atms: Sized + Debug {
    type PreCheckedPK: Copy;
    type PreparedPk;
    type SIG;
    fn prepare_keys(
        keys_pop: &[(Self::PreCheckedPK, Stake)]
    ) -> Option<(Vec<(Self::PreparedPk, Stake)>, Stake)>;

    fn verify(
        msg: &[u8], 
        aggregate_key: Self::PreparedPk, 
        nonsigners_aggregate_key: Self::PreparedPk, 
        sig: &Self::SIG
    ) -> bool;
}

The implementation relies on an underlying public signature with a proof of possession. We need the ability to verify the signature. Moreover, it is common that multi signatures schemes require the verification of correctness of the public key, and therefore we require also associate a check_key function.

Associated Types

type PreCheckedPK: Copy

Public key before verification of correctness.

type PreparedPk

Public key after preparation for ATMS protocol. This preparation may be a verification of a proof of correctness or a computation of an associated public exponent.

type SIG

Associated type that defines the Signature.

Required methods

fn prepare_keys(
keys_pop: &[(Self::PreCheckedPK, Stake)]
) -> Option<(Vec<(Self::PreparedPk, Stake)>, Stake)>

Prepare a set of public keys. If all are valid, return a vector formed of tuples of public keys and their corresponding stake, as well as the total stake.

fn verify(
msg: &[u8],
aggregate_key: Self::PreparedPk,
nonsigners_aggregate_key: Self::PreparedPk,
sig: &Self::SIG
) -> bool

Verify that the signature is valid for a message msg and key key