Trait mithril::atms::Atms

pub trait Atms: Sized + Debug {
    type POP: Copy;
    type PK;
    type SIG;
    fn verify_proof(proof: &Self::POP) -> Option<Self::PK>;

    fn verify(msg: &[u8], pk: &Self::PK, sig: &Self::SIG) -> bool;
}

The implementation relies on an underlying public signature with a proof of possession. We need the ability to verify the Proof of Possesion, and to verify the signature.

Associated Types

type POP: Copy

Proof of Possession of the secret key with respect to the Public Key.

type PK

Associated type that defines the Public Key.

type SIG

Associated type that defines the Signature.

Required methods

fn verify_proof(proof: &Self::POP) -> Option<Self::PK>

Verify the proof of possesion. If it is valid, return the public key (without the proof of possession)

fn verify(msg: &[u8], pk: &Self::PK, sig: &Self::SIG) -> bool

Verify that the signature is valid for a message msg and key key