Function mithril::stm::ev_lt_phi

pub fn ev_lt_phi(
    phi_f: f64, 
    ev: [u8; 64], 
    stake: Stake, 
    total_stake: Stake
) -> bool

Checks that ev is successful in the lottery. In particular, it compares the output of phi (a real) to the output of ev (a hash). It uses the same technique used in the Cardano ledger. In particular, ev is a natural in [0,2^512], while phi is a floating point in [0, 1], and so what this check does is verify whether p < 1 - (1 - phi_f)^w, with p = ev / 2^512.

The calculation is done using the following optimization:

let q = 1 / (1 - p) and c = ln(1 - phi_f)

then p < 1 - (1 - phi_f)^w <=> 1 / (1 - p) < exp(-w * c) <=> q < exp(-w * c)

This can be computed using the taylor expansion. Using error estimation, we can do an early stop, once we know that the result is either above or below. We iterate 1000 times. If no conclusive result has been reached, we return false.

Note that 1 1 evMax q = —– = —————— = ———–– 1 - p 1 - (ev / evMax) (evMax - ev)

Used to determine winning lottery tickets.